add invitedBy to member query
