+++ packages/ui/src/memberships/components/MemberReferrer/index.ts
@@ -1,1 +1,1 @@
-export * from './MemberReferre'
+export * from './MemberReferrer'
